move functions to better reflect original source
